refactor: move announcer struct declarations to announcer-common.h to reduce scope
